Solution to Check Number Sign

+++ Check Number Sign@@6788dd63490637cfc44e1b9c/code.c
@@ -2,8 +2,8 @@
 
 
 int main() {
-    float num;
-    scanf("%f",&num);
+    int num;
+    scanf("%d",&num);
     if(num>0){
         printf("Positive");
     }
